requirements.txt
selenium
HTMLTestRunner-rv